requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.32.0
PyYAML>=6.0.2
python-dotenv>=1.0.1
pydantic>=2.8.2
